lexiai/backend/requirements.txt
Flask>=2.2.0
flask-cors>=4.0.0
openai>=1.0.0
google-cloud-language>=2.0.0
huggingface-hub>=0.16.4,<1.0
transformers>=4.30.0
SQLAlchemy>=2.0.0
Jinja2>=3.1.0
python-dotenv>=1.0.0
requests>=2.30.0